fix the fileSize

+++ app/src/main/java/com/thkox/homeai/domain/models/Document.kt
@@ -3,6 +3,6 @@ package com.thkox.homeai.domain.models
 data class Document(
     val id: String,
     val fileName: String,
-    val fileSize: Float,
+    val fileSize: Int,
     val uploadTime: String
 )
