+++ app/src/main/java/com/thkox/homeai/presentation/models/DocumentUIModel.kt
@@ -3,6 +3,6 @@ package com.thkox.homeai.presentation.models
 data class DocumentUIModel(
     val id: String,
     val fileName: String,
-    val fileSize: Float,
+    val fileSize: Int,
     val uploadTime: String
 )
